Commit before merge

+++ SuperSim/Store.java
@@ -197,7 +197,7 @@ public class Store
 
     public int rebalanceCustomers() {
         //Assuming the the chances are going to vary to the nearest hour
-        int[] timeProbabilities =  {0.5, 0.5, 0.5, 0.5, 0.5, 0.5, 0.5, 0.5, 0.5, 0.5, 0.5, 0.5, 0.5, 0.5, 0.5, 0.5, 0.5, 0.5, 0.5, 0.5, 0.5, 0.5, 0.5, 0.5};
+        double[] timeProbabilities =  {0.5, 0.5, 0.5, 0.5, 0.5, 0.5, 0.5, 0.5, 0.5, 0.5, 0.5, 0.5, 0.5, 0.5, 0.5, 0.5, 0.5, 0.5, 0.5, 0.5, 0.5, 0.5, 0.5, 0.5};
         int currentHour = currentTick / 3600;
         int currentProbability = timeProbabilities[currentHour];
